==> Scripts/1 Data Wrangling & Subsetting.ipynb <==
# 1 Data Wrangling, Subsetting & Analysis

## Table of contents
1. [Importing Libraries & Data](#imports)
2. [Data Wrangling](#paragraph1)
3. [Data dictionary](#paragraph2)
4. [Subsetting](#paragraph3)
5. [Analysis](#paragraph4)
6. [Data export](#paragraph5)

### 1. Importing Libraries & Data <a name="imports"></a>

# Importing pandas, numpy and os libraries
import pandas as pd
import numpy as np
import os

# Turning project folder path into a string
path = r'C:\Users\deuts\Desktop\Instacart Basket Analysis'

# Importing 'orders.csv' 
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'),  index_col = False)

# Importing 'products.csv'
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

### 2. Data wrangling <a name="paragraph1"></a>

# Droping 'eval_set' column
df_ords = df_ords.drop(columns = ['eval_set'])

# Checking output
df_ords.head()

# Displaying entries for 'days_since_prior_order'
df_ords['days_since_prior_order'].value_counts(dropna = False)

# Renaming column 'order_dow' to 'orders_day_of_week' and checking output
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)
df_ords

# Displaying descriptive statistics of data frame
df_ords.describe()

# Changing data type of 'order_id' to sting
df_ords['order_id'] = df_ords['order_id'].astype('str')

# Importing new data frame 'departments.csv' and checking output
df_deps = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)
df_deps.head()

# Transposing index and columns of data frame
df_deps.T

# Overwriting data frame with transposed data frame and checking output
df_deps_t = df_deps.T
df_deps.T

# Resetting index of data frame
df_deps_t.reset_index()

# Creating new head for data frame and checking output
new_header = df_deps_t.iloc[0]
new_header

df_deps_t_new = df_deps_t[1:]

df_deps_t_new

df_deps_t_new.columns = new_header

new_header

df_deps_t_new

### 3. Data dictionary <a name="paragraph2"></a>

# Creating data dictionary
data_dict = df_deps_t_new.to_dict('index')
data_dict

print(data_dict.get('19'))

### 4. Subsetting <a name="paragraph3"></a>

# Creating new data frames for snacks and checking output
df_sacks = df_prods[df_prods['department_id'] == 19]
df_snacks

df_prods['department_id'] == 19

df_snacks = df_prods[df_prods['department_id'] == 19]

df_snacks.head()

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2

df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3

### 5. Analysis <a name="paragraph4"></a>

# Finding another idfentifier variable and changing it to a suitable format
## Displaying df_ords to identify idfentifiyer variable
df_ords

## Changing data type from integer to string and data type check of user_id
df_ords['user_id'] = df_ords['user_id'].astype('str')
df_ords['user_id'].dtype

# Looking for a variable in df_ords with an unituitive name
## Renaming user_id to 'customer_id' temporarily 
df_ords.rename(columns = {'user_id' : 'customer_id'}, inplace = False)

# Identifying busiest hour of the day
## Displaying descipitive statistics for order_hour_of_day
df_ords['order_hour_of_day'].describe()

## Changing data type from float to integer and data type check of order_hour_of_day
df_ords['order_hour_of_day'] = df_ords['order_hour_of_day'].astype('int')
df_ords['order_hour_of_day'].dtype

# Identifying busiest hour of the day
df_ords['order_hour_of_day'].value_counts(dropna = False)

##### Busiest hour of the day: 10 o'clock with 288418 orders

# 4. Alternative Identifying busiest hour of the day

# Changing data type from integer to string and data type check of order_hour_of_day
df_ords['order_hour_of_day'] = df_ords['order_hour_of_day'].astype('str')
df_ords['order_hour_of_day'].dtype

# Identifying busiest hour of the day
df_ords['order_hour_of_day'].describe()

# 5. Determing the meaning behind a value of 4 in 'department_id'
## Displaying departement for value 4
print(data_dict.get('4'))

##### department_id 4 = 'produce'

# 6. Creating subset for breakfast items
## Identifying department_id for 'breakfast'
df_deps_t_new.loc[df_deps_t_new['department'].isin(['breakfast'])]

## Filtering for breakfast items and displaying it
df_breakfast = df_prods[df_prods['department_id']==14]
df_breakfast

# Creating subset for dinner items
## Identifying department_ids for 'alcohol', 'deli', 'beverages', and 'meat/seafood'
df_deps_t_new.loc[df_deps_t_new['department'].isin(['alcohol', 'deli', 'beverages', 'meat seafood'])]

## Filtering for dinner items and displaying it
df_dinner = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]
df_dinner

##### 8. df_dinner contains 7650 rows

# 9. Extracting information of user_id = 1
## Filtering df_ords by user_id = 1

df_ords[df_ords['user_id']=='1']

## Displaying descriptive statistics on df_ords filtered on user_id = 1
df_ords[df_ords['user_id']=='1'].describe()

### 6. Data Export <a name="paragraph5"></a>

# Exporting updated orders data frame
df_ords.to_csv('orders_wrangled.csv')

# Exporting updated departments data frame
df_deps_t_new.to_csv('departments_wrangled.csv')
