==> Scripts/2 Data Consistency Checks.ipynb <==
# 2 Data Consistency Checks

## Table of contents
1. [Importing Libraries & Data](#imports)
2. [Products data frame](#paragraph1)
    1. [Mixed-type data](#subparagraph1)
    2. [Missing values](#subparagraph2)
    3. [Duplicates](#subparagraph3)
    4. [Export](#subparagraph3)
3. [Orders data frame](#paragraph2)
    1. [Mixed-type data](#subparagraph4)
    2. [Missing values](#subparagraph5)
    3. [Duplicates](#subparagraph6)
    4. [Export](#subparagraph7)

### 1. Importing Libraries & Data <a name="imports"></a>

# Importing pandas, numpy and os libraries
import pandas as pd
import numpy as np
import os

# Turning project folder path into a string
path = r'C:\Users\deuts\Desktop\Instacart Basket Analysis'

# Creating list with relevant 'orders' columns
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# Importing 'orders.csv' 
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# Importing 'products.csv'
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)


### 2. Products data frame <a name="paragraph1"></a>

### 2.1 Mixed-type data<a name="subparagraph1"></a>

# Create a data frame
df_test = pd.DataFrame()

# Create a mixed type columns and display it
df_test['mix'] = ['a','b',1,True]
df_test

# Check for mixed types in the test data frame
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Changing data type of 'mix' column to string
df_test['mix'] = df_test['mix'].astype('str')
df_test['mix'].dtype

### 2.2 Missing values <a name="subparagraph2"></a>

# Identifying missing values
df_prods.isnull().sum()

# Creating data frame with missing values
df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

# Checking the rows count for the products data frame
df_prods.shape

# Creating a new data frame for the products data frame
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Checking the rows count for the clean products data frame
df_prods_clean.shape

### 2.3 Duplicates <a name="subparagraph3"></a>

## Creating new data frame for duplicates of products data frame and displaying it to check the amount of duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

## Creating new clean data frame for clean products data frame and displaying it as a check
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups

### 2.4 Export <a name="subparagraph4"></a>

# Exporting clean products data frame
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_cleaned.csv'))

### 3. Orders data frame <a name="paragraph2"></a>

### 3.1 Mixed-type data <a name="subparagraph4"></a>

# 2. First look at the products data frame
## Descriptive Statistics
df_prods.describe()

## First five rows of the products data frame
df_prods.head()

#### The maximum seems off, as its a really high value, it seems like it is a cap.

# 3. Checking for mixed-type data
## Checking df_prods for mixed-type data
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

#### 4. No mixed-type data in the orders data frame

# 5. Missing data Check
## Running a check for missing data in the orders data frame
df_ords.isnull().sum()

### 3.2 Missing values <a name="subparagraph5"></a>

## Creating Sub data frame with missing data
df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]
df_ords_nan

#### For 206209 rows there is a null value/missing data in the 'days_since_prior_order' column. Explanation: For new customers it is not possible to have a previous order so the column 'days_since_prior_order' will have null value /missing data. 

# 6. Adressing the missing values
## Creating new data frame with a new column for new buyers
df_ords_clean = df_ords
df_ords_clean['new_customer'] = df_ords['days_since_prior_order'].isnull() == True
df_ords_clean

#### With the new colum 'new_customer' you can directly match a null value in the 'days_since_prior_order' to the customer type. So the  null value in the 'days_since_prior_order' column makes since, when the customer type is new customer (new_customer = True).

### 3.3 Duplicates <a name="subparagraph6"></a>

# 7. Check for duplicates
## Checking for duplicates in the orders data frame
df_ords_dups = df_ords_clean[df_ords_clean.duplicated()]
df_ords_dups

#### There are no duplicates in the orders data frame.

### 3.4 Export <a name="subparagraph7"></a>

# Exporting clean orders data frame
df_ords_clean.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_cleaned.csv'))

# Exporting clean products data frame
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_cleaned.csv'))
