==> Scripts/3 Combining & Exporting Data.ipynb <==
# 3 Combining & Exporting Data

## Table of contents
1. [Importing Libraries & Data](#imports)
2. [orders_products_prior data frame](#paragraph1)
    1. [Merge](#subparagraph1)
    2. [Export](#subparagraph2)
3. [orders_products_combined data frame](#paragraph2)
    1. [Import](#subparagraph3)
    2. [Merge](#subparagraph4)
    3. [Export](#subparagraph4)

### 1. Importing Libraries & Data <a name="imports"></a>

# Importing pandas, numpy and os libraries
import pandas as pd
import numpy as np
import os

# Turning project folder path into a string
path = r'C:\Users\deuts\Desktop\Instacart Basket Analysis'

# Importing 'orders_products_combined.csv'
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'order_products__prior.csv'))

# Importing 'orders.csv'
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'))

# Displaying column names of df_ords_prior
df_ords_prior.head()

# Displaying column names of df_ords
df_ords.head()

df_ords_prior.shape

df_ords.shape

### 2. orders_products_prior data frame <a name="paragraph1"></a>

### 2.1 Merge <a name="subparagraph1"></a>

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

### 2.2 Export <a name="subparagraph2"></a>

# Exporting data to csv
df_merged_large.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.csv'))
# Exporting data to pkl
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))

### 3. orders_products_combined data frame <a name="paragraph2"></a>

### 3.1 Import <a name="subparagraph2"></a>

# Importing 'orders_products_combined.csv'
orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# Checking columns of orders_products_combined.pkl
orders_products_combined.head()

# Checking shape of orders_products_combined.pkl
orders_products_combined.shape

# Import products.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_cleaned.csv'))
df_prods.head()

# Checking shape of products data frame
df_prods.shape

# Removing '_merge' column 
orders_products_combined = orders_products_combined.drop(columns = ['_merge'])

### 3.2 Merge <a name="subparagraph3"></a>

# Merging products and orders data frames
df_orders_products_merged = orders_products_combined.merge(df_prods, on = 'product_id', indicator = True)

# Displaying Columns of merged orders and products data frame 
df_orders_products_merged.head()

# Checking merge flag of merged data frame
df_orders_products_merged['_merge'].value_counts()

# Checking shape of merged data frame
df_orders_products_merged.shape

### 3.3 Export <a name="subparagraph4"></a>

# Export data to pkl
df_orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
