==> Scripts/4 Deriving New Variables.ipynb <==
# 4 Deriving New Variables

## Table of contents
1. [Importing Libraries & Data](#imports)
2. [Price Labels](#paragraph1)
3. [Busyness label](#paragraph2)
4. [Busiest days label](#paragraph3)
5. [Busiest period of day label](#paragraph4)
6. [Data export](#paragraph5)

### 1. Importing Libraries & Data <a name="imports"></a>

# Importing pandas, numpy and os libraries
import pandas as pd
import numpy as np
import os

# Turning project folder path into a string
path = r'C:\Users\deuts\Desktop\Instacart Basket Analysis'

# Importing 'orders_products_merged.pkl'
orders_products_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

orders_products_merged.head()

df = orders_products_merged[:1000000]

df.shape

### 2. Price labels <a name="paragraph1"></a>

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

orders_products_merged.loc[orders_products_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

orders_products_merged.loc[(orders_products_merged['prices'] <= 15) & (orders_products_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

orders_products_merged.loc[orders_products_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

orders_products_merged['price_range_loc'].value_counts(dropna = True)

# Renaming column from 'order_dow' to 'orders_day_of_the_week' again since I didnt't save it properly apperantly 
orders_products_merged = orders_products_merged.rename(columns={'order_dow': 'orders_day_of_week'})
orders_products_merged.head()

### 3. Busyness label <a name="paragraph2"></a>

orders_products_merged['orders_day_of_week'].value_counts(dropna = False)

result = []

for value in orders_products_merged['orders_day_of_week']:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

orders_products_merged['busiest_day'] = result

orders_products_merged['busiest_day'].value_counts(dropna = False)

### 4. Busiest days label <a name="paragraph3"></a>

## 2. Busiest days
# Creating for loop for checking the busiest days of the week
result_2 = []

for value in orders_products_merged['orders_day_of_week']:
  if value in [0,1]:
    result_2.append("Busiest day")
  elif value in [3,4]:
    result_2.append("Least busy")
  else:
    result_2.append("Regularly busy")

# Creating new column 'busiest_days' with the results of the for loop
orders_products_merged['busiest_days'] = result_2
orders_products_merged.head()

# Counting values of new column, sorted by business of days
orders_products_merged['busiest_days'].value_counts(dropna = False)

# 3. Checking result against output of business per day by the sum of the according single days
orders_products_merged['orders_day_of_week'].value_counts(dropna = False)

###### The sum of the days 0 (Saturday) & 1 (sunday) as beeing the busiest days is 6204182 + 5660230 = 11864412 equals the same as in the displayed data per day.
###### The sum of the days 3 (tuesday) & 4 (Wednesday) as beeing the slowest days is 3840534 + 3783802 = 7624336 equals the same as in the displayed data data per day.

### 5. Busiest period of day label <a name="paragraph4"></a>

## 4. Busiest hours
# Displaying orders per hour descending by orders
orders_products_merged['order_hour_of_day'].value_counts(dropna = False)

# Creating for loop for sorting the busiest hours of the day by periods
result_3 = []

for value in orders_products_merged['order_hour_of_day']:
  if value>=9 and value<=16:
    result_3.append("Most orders")
  elif value<=6:
    result_3.append("Fewest orders")
  else:
    result_3.append("Average orders")

# Creating new column 'busiest_period_of_day' with the results of the for loop
orders_products_merged['busiest_period_of_day'] = result_3
orders_products_merged.head()

# Counting values of new column by determined periods
orders_products_merged['busiest_period_of_day'].value_counts(dropna = False)

### 6. Data export <a name="paragraph5"></a>

## 7. Export
# Exporting data frame as pickle file to prepared data folder
orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_updated.pkl'))
