==> Scripts/5 Grouping Data & Aggregating Variables.ipynb <==
# 5 Grouping Data & Aggregating Variables

## Table of contents
1. [Importing Libraries & Data](#imports)
2. [Aggregated mean of order number](#paragraph1)
3. [Loyalty flag](#paragraph2)
4. [Spending habits](#paragraph3)
5. [Types of spenders](#paragraph4)
6. [Frequent versus non-frequent customers](#paragraph5)
7. [Data export](#paragraph6)

### 1. Importing Libraries & Data <a name="imports"></a>

# Importing pandas, numpy and os libraries
import pandas as pd
import numpy as np
import os

# Turning project folder path into a string
path = r'C:\Users\deuts\Desktop\Instacart Basket Analysis'

# Importing 'orders_products_merged_updated.pkl'
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_updated.pkl'))

### 2. Aggregated mean of order number <a name="paragraph1"></a>

# Checking shape of the data frame
ords_prods_merge.shape

# Displaying average amount of orders per department
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

# Caping a new data frame to 1000000 rows to compare to full data frame
df = ords_prods_merge[:1000000]

# Displaying average amount of orders per department for capped data frame
df.groupby('department_id').agg({'order_number': ['mean']})

##### The averages do not add up, since it's a larger data set now. The averages differ by up to 3 average orders.

### 3. Loyalty flag <a name="paragraph2"></a>

# Displaying basic statistics for amount of orders per department
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

# Creating new column 'max_order'
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# Checking output
ords_prods_merge.head(15)

# Creating loyalty flag for the whole df - 'Loyal customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

# Creating loyalty flag for the whole df - 'Regular customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

# Creating loyalty flag for the whole df - 'New customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Displaying output
ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

# Checking result
ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

### 4. Spending habits <a name="paragraph3"></a>

# Determining average, minumum and maximum of prices of products purchased by customer type
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

#### On average the New Customers spend the most in their orders, followed by regular and then loyal customers. The minimum with 1.0 and the maximum with 9999.0 spent is the same for alle Customer Types.

### 5. Types of spenders <a name="paragraph4"></a>

# Creating new column 'average_spending' and displaying output
ords_prods_merge['average_spending'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)
ords_prods_merge.head(15)

# Creating average_spending flag - 'Low spender'
ords_prods_merge.loc[ords_prods_merge['average_spending'] < 10, 'spending_flag'] = 'Low Spender'

# Creating loyalty flag - 'High spender '
ords_prods_merge.loc[(ords_prods_merge['average_spending'] >= 10), 'spending_flag'] = 'High Spender'

# Displaying output
ords_prods_merge['spending_flag'].value_counts(dropna = False)

### 6. Frequent versus non-frequent customers <a name="paragraph5"></a>

# Creating new column 'average_spending' and displaying output
ords_prods_merge['purchase_frequency'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)
ords_prods_merge.head(15)

# Creating frequency flag - 'Non-frequent customer'
ords_prods_merge.loc[ords_prods_merge['purchase_frequency'] > 20, 'frequency_flag'] = 'Non-frequent customer'

# Creating frequency flag - 'Regular customer'
ords_prods_merge.loc[(ords_prods_merge['purchase_frequency'] > 10) & (ords_prods_merge['purchase_frequency'] <= 20), 'frequency_flag'] = 'Regular customer'

# Creating frequency flag - 'Frequent customer'
ords_prods_merge.loc[(ords_prods_merge['purchase_frequency'] <= 10), 'frequency_flag'] = 'Frequent Customer'

# Displaying output
ords_prods_merge['frequency_flag'].value_counts(dropna = False)

### 7. Data export <a name="paragraph6"></a>

# Exporting data frame as pickle file to prepared data folder
ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge_final.pkl'))
