==> Scripts/6 Final data wrangling, consistency check and merging.ipynb <==
# 6 Final data wrangling, consistency check and merging

## Table of contents
1. [Importing Libraries & Data](#imports)
2. [Data wrangling](#paragraph1)
3. [Data consistency & quality](#paragraph2)
4. [Data merge](#paragraph3)
5. [Data export](#paragraph4)

### 1. Importing Libraries & Data <a name="imports"></a>

# Importing pandas, numpy, os, matplotlib, seaborn & scipy libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# Turning project folder path into a string
path = r'C:\Users\deuts\Desktop\Instacart Basket Analysis'

# Importing 'customers.csv' as df_customers
df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

### 2. Data wrangling <a name="paragraph1"></a>

# Check shape of 'customers' data frame
df_customers.shape

# Check column names of 'customers' data frame
df_customers.head()

# Renaming column names
df_customers.rename(columns={'First Name':'first_name', 'Surnam':'last_name','Gender':'gender','STATE':'state','Age':'age','n_dependants':'no_of_dependants','fam_status':'relationship_status'}, inplace=True)

# Checking output
df_customers.head()

# Displaying basic statistics
df_customers.describe()

# Changing data type for column 'user_id', since it does not need calculated statistics and checking data type after change
df_customers['user_id'] = df_customers['user_id'].astype('str')
df_customers['user_id'].dtype

df_customers.describe()

### 3. Data consistency & quality <a name="paragraph2"></a>

df_customers.head()

df_customers.describe()

# Checking for mixed types data
for col in df_customers.columns.tolist():
  mixed_types = (df_customers[[col]].applymap(type) != df_customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customers[mixed_types]) > 0:
    print (col)

##### Column 'first name' has mixed types data in the customers data frame.

# Changing data type for column 'first_name', since a name has to be a string and checking data type after change
df_customers['first_name'] = df_customers['first_name'].astype('str')
df_customers['first_name'].dtype

# Checking for missing values
df_customers.isnull().sum()

##### No missing data in the customers data frame.

# Cheking for duplicates
df_customers_dups = df_customers[df_customers.duplicated()]
df_customers_dups

##### No duplicates in the customers data frame.

### 4. Data merge <a name="paragraph3"></a>

# Importing 'orders_products_merged_updated.pkl' and displaying head
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge_final.pkl'))
ords_prods_merge.head()

# Dropping not needed column 'Unnamed: 0' and checking output
ords_prods_merge = ords_prods_merge.drop('Unnamed: 0', axis=1)
ords_prods_merge.head()

# Checking shape of 'ords_prods_merge'
ords_prods_merge.shape

# Checking head of 'df_customers'
df_customers.head()

# Checking shape of 'ords_prods_merge'
df_customers.shape

#### Common column for merging the two data frames is 'user_id'. 

# Checking if data types of columns 'user_id' in data frame 'ords_prods_merge' and data frame 'df_customers' are the same
ords_prods_merge['user_id'].dtype == df_customers['user_id'].dtype

#### Data types of the 'user_id' columns in the two data frames are not the same.

# Checking data type of column 'user_id' in data frame 'ords_prods_merge'
ords_prods_merge['user_id'].dtype

# Checking data type of column 'user_id' in data frame 'df_customers'
df_customers['user_id'].dtype

# Rechanging data type for column 'user_id' (df_customers), since it has to be the same data type as in the other data frame, and integer is most suitable for 'user_id'
df_customers['user_id'] = df_customers['user_id'].astype('int64')
df_customers['user_id'].dtype

# Rechecking if data types of columns 'user_id' in data frame 'ords_prods_merge' and data frame 'df_customers' are the same
ords_prods_merge['user_id'].dtype == df_customers['user_id'].dtype

#### Data types of the 'user_id' columns in the two data frames are the same now.

# Merging both data frames
ords_prods_customers_merge = ords_prods_merge.merge(df_customers, on = 'user_id')

ords_prods_customers_merge.head()

ords_prods_customers_merge.shape

### 5. Data export <a name="paragraph4"></a>

# Exporting data frame as pickle file to prepared data folder
ords_prods_customers_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'df_instacart.pkl'))
